v0.3.0.1 - minimal cli version for public use

+++ src/vasco/cli_new.py
@@ -8,11 +8,10 @@ from typing_extensions import Annotated
 from vasco.config import vasco_data_dir
 
 from vasco.util import casadir_find, rfc_find
-from vasco.pipe.config import DEFAULT_PARAMS
+from vasco.pipe.config import CSV_POPULATED_STEPS, PipeConfig, DEFAULT_PARAMS
 
 from vasco.pipe.main import VascoPipeline
 from vasco.pipe.core import PipelineContext
-from vasco.pipe.config import PipeConfig
 
 rlimit = resource.getrlimit(resource.RLIMIT_NOFILE)
 resource.setrlimit(resource.RLIMIT_NOFILE, (26000, rlimit[1]))
@@ -65,12 +64,12 @@ def setup_rfc(rfc_filepath):
     """Set the RFC calibrator list path."""
     rfc_find(rfc_filepath, write=True)
 
-# __________________________
+# __________________________    without command
 
-listobs_data = typer.Typer(help="List observation data")
-vasco_cli.add_typer(listobs_data, name="listobs")
+listobs_app = typer.Typer(help="List observation data")
+vasco_cli.add_typer(listobs_app, name="listobs")
 
-@listobs_data.callback(invoke_without_command=True)
+@listobs_app.callback(invoke_without_command=True)
 def listobs(fitsfilenames: Annotated[Optional[List[str]], typer.Argument()] = None):
     from vasco.fitsidiutil import ObservationSummary
     print(ObservationSummary(fitsfilepaths=fitsfilenames).to_polars())
@@ -79,21 +78,39 @@ def listobs(fitsfilenames: Annotated[Optional[List[str]], typer.Argument()] = No
     # print(df_obsdata)
 
 
+fitsidicheck_app = typer.Typer(help="validate and fix, known FITS-IDI problems")
+vasco_cli.add_typer(fitsidicheck_app, name="validate")
+
+@fitsidicheck_app.callback(invoke_without_command=True)
+def fitsidicheck(fitsfilenames: Annotated[Optional[List[str]], typer.Argument()] = None,
+                 fix:bool=False, desc:bool=False):
+    """
+    "validate and fix, known FITS-IDI problems"
+    """
+    from vasco.fitsidiutil.validation import fitsidi_check
+    for fitsfile in fitsfilenames:
+        validators = fitsidi_check(fitsfilepath=fitsfile)
+        if desc:
+            print(validators)
+        else:
+            print(validators.run(fix=fix))
+        
+    
+
 
 
 # ___________________________
 
 
 pipeline_app = typer.Typer(help="VASCO pipeline.")
-vasco_cli.add_typer(pipeline_app, name="pipeline")
+vasco_cli.add_typer(pipeline_app, name="pipe")
 
 @pipeline_app.command("run")
 def run_pipeline(
-    # filter_steps: str = typer.Option("", help="Comma-separated list of steps to run"),
-    # fitsfilenames: str = typer.Option("", help="Comma-separated list of fitsfile names"),
-    fitsfilenames: Annotated[Optional[List[str]],typer.Argument(help="Input file path(s).")] = None, 
-    target: str = typer.Option("", help="target name"),
-    configfile: Optional[str] = typer.Option(None, help="config file"),
+    fitsfilenames: Annotated[str,typer.Option("--f", "--fitsfilenames", help="fitsfile names comma separated")] = '',
+    stps: Annotated[Optional[List[str]],typer.Argument(help="steps for execution")] = CSV_POPULATED_STEPS, 
+    target: Annotated[str,typer.Option("--t", "--target", help="Selected field / sourc name")] = '',
+    configfile: Optional[str] = typer.Option("vasco.inp", help="config file containing key=value"),
     ):
     """
     _______________________
@@ -118,19 +135,24 @@ def run_pipeline(
                 "folder_for_fits": ".",
                  "target_dir" : "reduction/", 
                  "primary_value": target, 
-                 "casadir":"/home/avi/intelligence/env/casa-6.7.0-31-py3.10.el8/",
-                 "rfc_catalogfile":"rfc_2024a_cat.txt", 
+                #  "casadir":"/home/avi/intelligence/env/casa-6.7.0-31-py3.10.el8/",
+                #  "rfc_catalogfile":"rfc_2024a_cat.txt", 
                  "target":target,
-                 "fitsfilenames": fitsfilenames,
+                 "fitsfilenames": fitsfilenames.split(","),
                  }    
-
+    
+    
+    pipe_params.update(PipeConfig(configfile).to_dict())
+    
+    # if configfile:
+    #     configdata = PipeConfig(configfile=configfile)
+    #     pipe_params.update(configdata.to_dict())
+    
+    # print(DEFAULT_PARAMS['allfitsfile'])
     main_pipeline = VascoPipeline(pipe_params=pipe_params)
     
-    if configfile:
-        configdata = PipeConfig(configfile=configfile)
-        pipe_params.update(configdata.to_dict())
-    
-    
+    # main_pipeline.execute()
+    main_pipeline.filter_steps(*stps)
     result = main_pipeline.execute()
     
     target = f"{pipe_params['target']}_"
@@ -138,8 +160,12 @@ def run_pipeline(
     
     
     print(result)
-    result.to_polars().write_csv(csvfile)
-
+    
+    if not Path(csvfile).exists():
+        result.to_polars().write_csv(csvfile)
+    else:
+        with open(csvfile, "ab") as f:
+            result.to_polars().write_csv(f, include_header=False)
 
     
 if __name__=='__main__':
